requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
sqlparse==0.4.1
whitenoise==5.2.0

